Backend done

+++ order/src/main/java/com/capgemini/go/order/service/OrderServiceImpl.java
@@ -1,13 +1,15 @@
 package com.capgemini.go.order.service;
 
+import java.time.LocalDate;
 import java.util.List;
+import java.util.Random;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import com.capgemini.go.order.exception.CancelException;
 import com.capgemini.go.order.exception.OrderIdNotFoundException;
+import com.capgemini.go.order.exception.ProductNotFoundException;
 import com.capgemini.go.order.model.OrderDto;
 import com.capgemini.go.order.repository.IOrderRepo;
-import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class OrderServiceImpl implements IOrderService {
@@ -15,21 +17,27 @@ public class OrderServiceImpl implements IOrderService {
 	@Autowired
 	IOrderRepo orderRepository;
 	
+	@Autowired
+	private Random random;
+	
+	LocalDate date=LocalDate.now();
+	LocalDate dispatchDate=date.plusDays(2);
+	LocalDate cantCancel=dispatchDate.MAX;
+	
 	@Override
 	public OrderDto addOrder(OrderDto orderDto) {
 		
+		orderDto.setProductUniqueNo(random.nextInt(1000));
 		return orderRepository.save(orderDto);
 	}
 
 	@Override
 	public OrderDto viewOrder(String orderId) {
-		
+	
 		if(!orderRepository.existsById(orderId)) {
 			throw new OrderIdNotFoundException("The given OrderId is not present");
 		}
-		else if(orderId==null) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-		}
+		
 		return orderRepository.getOne(orderId);
 	}
 
@@ -40,11 +48,32 @@ public class OrderServiceImpl implements IOrderService {
 	}
 
 	@Override
-	public void deleteOrder(String orderId) {
+	public void cancelOrder(String orderId) {
 		
 		if(!orderRepository.existsById(orderId)) {
 			throw new OrderIdNotFoundException("The given OrderId is not present");
 		}
+		if((dispatchDate.compareTo(cantCancel))>0 || (dispatchDate.compareTo(cantCancel)==0)) {
+			throw new CancelException("You cannot cancel your order because it was already dispatched");
+		}
 		orderRepository.deleteById(orderId); 
+	}
+
+	@Override
+	public void cancelProduct(String orderId,String productId) {
+
+		if(!orderRepository.existsById(orderId)) {
+			throw new OrderIdNotFoundException("The given OrderId or ProductId is not present");
+		}
+		if(!orderRepository.findAll().contains(orderRepository.existsProductId(productId))) {
+			throw new ProductNotFoundException("The given ProductId is not valid");
+		}				
+		orderRepository.deleteById(productId);
+	}
+
+	@Override
+	public String dispatchDate() {
+		
+		return "You have ordered your item on "+ date +". So "+ dispatchDate + " will be your Dispatch date";
 	}
 }
